+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADConditionalAccessPolicy/MSFT_AADConditionalAccessPolicy.psm1
@@ -27,6 +27,15 @@ function Get-TargetResource
         $ExcludeApplications,
 
         [Parameter()]
+        [System.String]
+        $ApplicationsFilter,
+
+        [Parameter()]
+        [ValidateSet("include", "exclude")]
+        [System.String]
+        $ApplicationsFilterMode,
+
+        [Parameter()]
         [System.String[]]
         $IncludeUserActions,
 
@@ -397,7 +406,6 @@ function Get-TargetResource
         }
     }
 
-
     $IncludeRoles = @()
     $ExcludeRoles = @()
     #translate role template guids to role name
@@ -450,7 +458,6 @@ function Get-TargetResource
                 }
             }
         }
-
     }
 
     $IncludeLocations = @()
@@ -600,6 +607,8 @@ function Get-TargetResource
         IncludeApplications                      = [System.String[]](@() + $Policy.Conditions.Applications.IncludeApplications)
         #no translation of Application GUIDs, return empty string array if undefined
         ExcludeApplications                      = [System.String[]](@() + $Policy.Conditions.Applications.ExcludeApplications)
+        ApplicationsFilter                       = $Policy.Conditions.Applications.ApplicationFilter.Rule
+        ApplicationsFilterMode                   = $Policy.Conditions.Applications.ApplicationFilter.Mode
         #no translation of GUIDs, return empty string array if undefined
         IncludeUserActions                       = [System.String[]](@() + $Policy.Conditions.Applications.IncludeUserActions)
         #no translation needed, return empty string array if undefined
@@ -640,20 +649,20 @@ function Get-TargetResource
         BuiltInControls                          = [System.String[]](@() + $Policy.GrantControls.BuiltInControls)
         CustomAuthenticationFactors              = [System.String[]](@() + $Policy.GrantControls.CustomAuthenticationFactors)
         #no translation needed, return empty string array if undefined
-        ApplicationEnforcedRestrictionsIsEnabled = $false -or $Policy.SessionControls.ApplicationEnforcedRestrictions.IsEnabled
+        ApplicationEnforcedRestrictionsIsEnabled = $Policy.SessionControls.ApplicationEnforcedRestrictions.IsEnabled
         #make false if undefined, true if true
-        CloudAppSecurityIsEnabled                = $false -or $Policy.SessionControls.CloudAppSecurity.IsEnabled
+        CloudAppSecurityIsEnabled                = $Policy.SessionControls.CloudAppSecurity.IsEnabled
         #make false if undefined, true if true
         CloudAppSecurityType                     = [System.String]$Policy.SessionControls.CloudAppSecurity.CloudAppSecurityType
         #no translation needed, return empty string array if undefined
-        SignInFrequencyIsEnabled                 = $false -or $Policy.SessionControls.SignInFrequency.IsEnabled
+        SignInFrequencyIsEnabled                 = $Policy.SessionControls.SignInFrequency.IsEnabled
         #make false if undefined, true if true
         SignInFrequencyValue                     = $Policy.SessionControls.SignInFrequency.Value
         #no translation or conversion needed, $null returned if undefined
         SignInFrequencyType                      = [System.String]$Policy.SessionControls.SignInFrequency.Type
         SignInFrequencyInterval                  = $SignInFrequencyIntervalValue
         #no translation needed
-        PersistentBrowserIsEnabled               = $false -or $Policy.SessionControls.PersistentBrowser.IsEnabled
+        PersistentBrowserIsEnabled               = $Policy.SessionControls.PersistentBrowser.IsEnabled
         #make false if undefined, true if true
         PersistentBrowserMode                    = [System.String]$Policy.SessionControls.PersistentBrowser.Mode
         #no translation needed
@@ -700,6 +709,15 @@ function Set-TargetResource
         [Parameter()]
         [System.String[]]
         $ExcludeApplications,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationsFilter,
+
+        [Parameter()]
+        [ValidateSet("include", "exclude")]
+        [System.String]
+        $ApplicationsFilterMode,
 
         [Parameter()]
         [System.String[]]
@@ -940,10 +958,19 @@ function Set-TargetResource
         {
             $conditions.Applications.Add('IncludeApplications', $IncludeApplications)
         }
-        if ($ExcludeApplications)
+        if ($currentParameters.ContainsKey("ExcludeApplications"))
         {
             $conditions.Applications.Add('ExcludeApplications', $ExcludeApplications)
         }
+        if ($ApplicationsFilter -and $ApplicationsFilterMode)
+        {
+            $appFilterValue = @{
+                rule = $ApplicationsFilter
+                mode = $ApplicationsFilterMode
+            }
+            $conditions.Applications.Add("ApplicationFilter", $appFilterValue)
+        }
+
         if ($IncludeUserActions)
         {
             $conditions.Applications.Add('IncludeUserActions', $IncludeUserActions)
@@ -1447,15 +1474,13 @@ function Set-TargetResource
         {
             Write-Verbose -Message 'Set-Targetresource: create provision Session Control object'
             $sessioncontrols = @{
-                ApplicationEnforcedRestrictions = @{
-                    IsEnabled = $false
-                }
+                ApplicationEnforcedRestrictions = @{}
             }
 
-            if ($ApplicationEnforcedRestrictionsIsEnabled)
+            if ($ApplicationEnforcedRestrictionsIsEnabled -eq $true)
             {
                 #create and provision ApplicationEnforcedRestrictions object if used
-                $sessioncontrols.ApplicationEnforcedRestrictions.IsEnabled = $true
+                $sessioncontrols.ApplicationEnforcedRestrictions.Add('IsEnabled', $true)
             }
             if ($CloudAppSecurityIsEnabled)
             {
@@ -1603,6 +1628,15 @@ function Test-TargetResource
         [Parameter()]
         [System.String[]]
         $ExcludeApplications,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationsFilter,
+
+        [Parameter()]
+        [ValidateSet("include", "exclude")]
+        [System.String]
+        $ApplicationsFilterMode,
 
         [Parameter()]
         [System.String[]]