+++ Modules/Microsoft365DSC/Examples/Resources/AADAuthenticationMethodPolicyX509/1-Create.ps1
@@ -18,10 +18,8 @@ Configuration Example
         AADAuthenticationMethodPolicyX509 "AADAuthenticationMethodPolicyX509-X509Certificate"
         {
             AuthenticationModeConfiguration = MSFT_MicrosoftGraphx509CertificateAuthenticationModeConfiguration{
-
-                Rules = @(@()
-                )
                 X509CertificateAuthenticationDefaultMode = 'x509CertificateSingleFactor'
+                Rules = @(@())
             };
             CertificateUserBindings         = @(
                 MSFT_MicrosoftGraphx509CertificateUserBinding{
@@ -44,7 +42,7 @@ Configuration Example
             Ensure                          = "Present";
             ExcludeTargets                  = @(
                 MSFT_AADAuthenticationMethodPolicyX509ExcludeTarget{
-                    Id = 'DSCGroup'
+                    Id = 'Sales Team'
                     TargetType = 'group'
                 }
             );