+++ Modules/Microsoft365DSC/Examples/Resources/AADConditionalAccessPolicy/2-Update.ps1
@@ -15,12 +15,10 @@ Configuration Example
 
     node localhost
     {
-        AADConditionalAccessPolicy 'Allin-example'
+        AADConditionalAccessPolicy 'ConditionalAccessPolicy'
         {
-            ApplicationEnforcedRestrictionsIsEnabled = $False;
             BuiltInControls                          = @("mfa");
             ClientAppTypes                           = @("all");
-            CloudAppSecurityIsEnabled                = $True; # Updated Porperty
             Credential                               = $Credscredential;
             DeviceFilterMode                         = "exclude";
             DeviceFilterRule                         = "device.trustType -eq `"AzureAD`" -or device.trustType -eq `"ServerAD`" -or device.trustType -eq `"Workplace`"";
@@ -30,11 +28,10 @@ Configuration Example
             GrantControlOperator                     = "OR";
             IncludeApplications                      = @("All");
             IncludeRoles                             = @("Attack Payload Author");
-            PersistentBrowserIsEnabled               = $False;
             SignInFrequencyInterval                  = "timeBased";
             SignInFrequencyIsEnabled                 = $True;
             SignInFrequencyType                      = "hours";
-            SignInFrequencyValue                     = 1;
+            SignInFrequencyValue                     = 2; # Updated Porperty
             State                                    = "disabled";
         }
     }